1. Implement Factory & Reflection design pattern for create instance based on calculator type. 2. Create & modified testcases based on factory class. 3. Enhanced income tax and payslip calculator. 4. Waiting for fill up monthly payslip calculator functional code.

+++ CodingExercise.IncomeTaxCalculator/Interface/ITaxIncomeCalculator.cs
@@ -11,6 +11,6 @@ namespace CodingExercise.IncomeTaxCalculator.Interface
     /// </summary>
     public interface ITaxIncomeCalculator
     {
-        int CalculateIncomeTax(int annualSalary);
+        double CalculateIncomeTax(int annualSalary);
     }
 }
